+++ buildscript-utils/src/main/kotlin/com/fueledbycaffeine/spotlight/buildscript/models/SpotlightRules.kt
@@ -6,6 +6,7 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 public data class SpotlightRules(
   val implicitRules: Set<ImplicitDependencyRule> = emptySet(),
+  val taskInvocationRules: Set<TaskInvocationRule> = emptySet(),
 ) {
   public companion object {
     public val EMPTY: SpotlightRules = SpotlightRules()